chore: info message
add info message for .well-known endpoints
